chroma-manager/requirements.txt
# Lock python in at 2.6 (EL6 is 2.6.6)
python>=2.6.5,<2.7.0
django==1.4.1
nose>=1.1.2
django-nose
django-extensions
django-celery==3.0.10
django-picklefield
kombu==2.4.7
celery==3.0.11
mock
paramiko
pep8==1.0.1
pyflakes
South==0.7.6
sphinx==1.1.3
werkzeug
MySQL-python
coverage
python-daemon==1.5.2
lockfile==0.8
django-tastypie==0.9.11
tablib
requests>=1.0.3
nose-testconfig
fudge
simplejson
behave>=1.1.0
prettytable
dse
selenium
log4tailer
jsonschema
supervisor
django-supervisor
gevent
m2crypto
pygraphviz # For building diagrams in the docs
wsgiproxy
